BACKEND/requirement.txt
# Core system monitoring
psutil>=7.0.0

# Weather API
requests>=2.32.0
python-dotenv>=1.0.0

# Voice and speech
sounddevice>=0.5.0
numpy>=2.0.0
speechrecognition>=3.10.0
pyttsx3>=2.90
pyaudio>=0.2.11  # For Windows/Mac, use: conda install -c conda-forge pyaudio on Linux

# AI and Ollama
ollama>=0.6.0

# Email (if using IMAP method)
imap-tools>=1.7.0

# For Gmail API (optional - uncomment if using)
# google-api-python-client>=2.0.0
# google-auth-oauthlib>=1.0.0
# google-auth-httplib2>=0.1.0
# google-generativeai>=0.8.0

# File operations (built-in but listed for completeness)
# os, json, datetime, getpass (built into Python)